requirements.txt

pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
transformers
nltk
joblib
sqlite-utils
openai
